Add LinearRegression 3.1.1 & 3.1.2
