+++ mllib-dal/src/main/scala/org/apache/spark/ml/regression/LinearRegression.scala
@@ -17,113 +17,14 @@
 
 package org.apache.spark.ml.regression
 
-import scala.collection.mutable
-import breeze.linalg.{DenseVector => BDV}
-import breeze.optimize.{CachedDiffFunction, DiffFunction, FirstOrderMinimizer, LBFGS => BreezeLBFGS, LBFGSB => BreezeLBFGSB, OWLQN => BreezeOWLQN}
-import breeze.stats.distributions.StudentsT
-import org.apache.hadoop.fs.Path
-import org.apache.spark.SparkException
+import com.intel.oap.mllib.regression.LinearRegressionShim
 import org.apache.spark.annotation.Since
 import org.apache.spark.internal.Logging
-import org.apache.spark.ml.{PipelineStage, PredictorParams}
-import org.apache.spark.ml.feature._
-import org.apache.spark.ml.linalg.{Vector, Vectors}
-import org.apache.spark.ml.linalg.BLAS._
+import org.apache.spark.ml.linalg.Vector
 import org.apache.spark.ml.optim.WeightedLeastSquares
-import org.apache.spark.ml.optim.aggregator._
-import org.apache.spark.ml.optim.loss.{L2Regularization, RDDLossFunction}
-import org.apache.spark.ml.param.{DoubleParam, Param, ParamMap, ParamValidators}
-import org.apache.spark.ml.param.shared._
-import org.apache.spark.ml.stat._
+import org.apache.spark.ml.param.ParamMap
 import org.apache.spark.ml.util._
-import org.apache.spark.ml.util.Instrumentation.instrumented
-import org.apache.spark.mllib.evaluation.RegressionMetrics
-import org.apache.spark.mllib.linalg.VectorImplicits._
-import org.apache.spark.mllib.regression.{LinearRegressionModel => OldLinearRegressionModel}
-import org.apache.spark.mllib.util.MLUtils
-import org.apache.spark.rdd.RDD
-import org.apache.spark.sql.{DataFrame, Dataset, Row, SparkSession}
-import org.apache.spark.sql.functions._
-import org.apache.spark.sql.types.{DataType, DoubleType, StructType}
-import org.apache.spark.storage.StorageLevel
-import org.apache.spark.util.VersionUtils.majorMinorVersion
-import com.intel.oap.mllib.Utils
-import com.intel.oap.mllib.regression.LinearRegressionDALImpl
-
-/**
- * Params for linear regression.
- */
-private[regression] trait LinearRegressionParams extends PredictorParams
-    with HasRegParam with HasElasticNetParam with HasMaxIter with HasTol
-    with HasFitIntercept with HasStandardization with HasWeightCol with HasSolver
-    with HasAggregationDepth with HasLoss with HasMaxBlockSizeInMB {
-
-  import LinearRegression._
-
-  /**
-   * The solver algorithm for optimization.
-   * Supported options: "l-bfgs", "normal" and "auto".
-   * Default: "auto"
-   *
-   * @group param
-   */
-  @Since("1.6.0")
-  final override val solver: Param[String] = new Param[String](this, "solver",
-    "The solver algorithm for optimization. Supported options: " +
-      s"${supportedSolvers.mkString(", ")}. (Default auto)",
-    ParamValidators.inArray[String](supportedSolvers))
-
-  /**
-   * The loss function to be optimized.
-   * Supported options: "squaredError" and "huber".
-   * Default: "squaredError"
-   *
-   * @group param
-   */
-  @Since("2.3.0")
-  final override val loss: Param[String] = new Param[String](this, "loss", "The loss function to" +
-    s" be optimized. Supported options: ${supportedLosses.mkString(", ")}. (Default squaredError)",
-    ParamValidators.inArray[String](supportedLosses))
-
-  /**
-   * The shape parameter to control the amount of robustness. Must be &gt; 1.0.
-   * At larger values of epsilon, the huber criterion becomes more similar to least squares
-   * regression; for small values of epsilon, the criterion is more similar to L1 regression.
-   * Default is 1.35 to get as much robustness as possible while retaining
-   * 95% statistical efficiency for normally distributed data. It matches sklearn
-   * HuberRegressor and is "M" from <a href="http://statweb.stanford.edu/~owen/reports/hhu.pdf">
-   * A robust hybrid of lasso and ridge regression</a>.
-   * Only valid when "loss" is "huber".
-   *
-   * @group expertParam
-   */
-  @Since("2.3.0")
-  final val epsilon = new DoubleParam(this, "epsilon", "The shape parameter to control the " +
-    "amount of robustness. Must be > 1.0.", ParamValidators.gt(1.0))
-
-  /** @group getExpertParam */
-  @Since("2.3.0")
-  def getEpsilon: Double = $(epsilon)
-
-  setDefault(regParam -> 0.0, fitIntercept -> true, standardization -> true,
-    elasticNetParam -> 0.0, maxIter -> 100, tol -> 1E-6, solver -> Auto,
-    aggregationDepth -> 2, loss -> SquaredError, epsilon -> 1.35, maxBlockSizeInMB -> 0.0)
-
-  override protected def validateAndTransformSchema(
-      schema: StructType,
-      fitting: Boolean,
-      featuresDataType: DataType): StructType = {
-    if (fitting) {
-      if ($(loss) == Huber) {
-        require($(solver)!= Normal, "LinearRegression with huber loss doesn't support " +
-          "normal solver, please change solver to auto or l-bfgs.")
-        require($(elasticNetParam) == 0.0, "LinearRegression with huber loss only supports " +
-          s"L2 regularization, but got elasticNetParam = $getElasticNetParam.")
-      }
-    }
-    super.validateAndTransformSchema(schema, fitting, featuresDataType)
-  }
-}
+import org.apache.spark.sql.Dataset
 
 /**
  * Linear regression.
@@ -185,8 +86,6 @@ private[regression] trait LinearRegressionParams extends PredictorParams
 class LinearRegression @Since("1.3.0") (@Since("1.3.0") override val uid: String)
   extends Regressor[Vector, LinearRegression, LinearRegressionModel]
   with LinearRegressionParams with DefaultParamsWritable with Logging {
-
-  import LinearRegression._
 
   @Since("1.4.0")
   def this() = this(Identifiable.randomUID("linReg"))
@@ -325,351 +224,10 @@ class LinearRegression @Since("1.3.0") (@Since("1.3.0") override val uid: String
   def setMaxBlockSizeInMB(value: Double): this.type = set(maxBlockSizeInMB, value)
 
   override protected def train(
-      dataset: Dataset[_]): LinearRegressionModel = instrumented { instr =>
-    instr.logPipelineStage(this)
-    instr.logDataset(dataset)
-    instr.logParams(this, labelCol, featuresCol, weightCol, predictionCol, solver, tol,
-      elasticNetParam, fitIntercept, maxIter, regParam, standardization, aggregationDepth, loss,
-      epsilon, maxBlockSizeInMB)
-
-    if (dataset.storageLevel != StorageLevel.NONE) {
-      instr.logWarning(s"Input instances will be standardized, blockified to blocks, and " +
-        s"then cached during training. Be careful of double caching!")
-    }
-
-    // Extract the number of features before deciding optimization solver.
-    val numFeatures = MetadataUtils.getNumFeatures(dataset, $(featuresCol))
-    instr.logNumFeatures(numFeatures)
-
-    if ($(loss) == SquaredError && (($(solver) == Auto &&
-      numFeatures <= WeightedLeastSquares.MAX_NUM_FEATURES) || $(solver) == Normal)) {
-      return trainWithNormal(dataset, instr)
-    }
-
-    val instances = extractInstances(dataset)
-      .setName("training instances")
-
-    val (summarizer, labelSummarizer) = Summarizer
-      .getRegressionSummarizers(instances, $(aggregationDepth), Seq("mean", "std", "count"))
-
-    val yMean = labelSummarizer.mean(0)
-    val rawYStd = labelSummarizer.std(0)
-
-    instr.logNumExamples(labelSummarizer.count)
-    instr.logNamedValue(Instrumentation.loggerTags.meanOfLabels, yMean)
-    instr.logNamedValue(Instrumentation.loggerTags.varianceOfLabels, rawYStd)
-    instr.logSumOfWeights(summarizer.weightSum)
-
-    var actualBlockSizeInMB = $(maxBlockSizeInMB)
-    if (actualBlockSizeInMB == 0) {
-      actualBlockSizeInMB = InstanceBlock.DefaultBlockSizeInMB
-      require(actualBlockSizeInMB > 0, "inferred actual BlockSizeInMB must > 0")
-      instr.logNamedValue("actualBlockSizeInMB", actualBlockSizeInMB.toString)
-    }
-
-    if (rawYStd == 0.0) {
-      if ($(fitIntercept) || yMean == 0.0) {
-        return trainWithConstantLabel(dataset, instr, numFeatures, yMean)
-      } else {
-        require($(regParam) == 0.0, "The standard deviation of the label is zero. " +
-          "Model cannot be regularized.")
-        instr.logWarning(s"The standard deviation of the label is zero. " +
-          "Consider setting fitIntercept=true.")
-      }
-    }
-
-    // if y is constant (rawYStd is zero), then y cannot be scaled. In this case
-    // setting yStd=abs(yMean) ensures that y is not scaled anymore in l-bfgs algorithm.
-    val yStd = if (rawYStd > 0) rawYStd else math.abs(yMean)
-    val featuresMean = summarizer.mean.toArray
-    val featuresStd = summarizer.std.toArray
-
-    if (!$(fitIntercept) && (0 until numFeatures).exists { i =>
-      featuresStd(i) == 0.0 && featuresMean(i) != 0.0 }) {
-      instr.logWarning("Fitting LinearRegressionModel without intercept on dataset with " +
-        "constant nonzero column, Spark MLlib outputs zero coefficients for constant nonzero " +
-        "columns. This behavior is the same as R glmnet but different from LIBSVM.")
-    }
-
-    // Since we implicitly do the feature scaling when we compute the cost function
-    // to improve the convergence, the effective regParam will be changed.
-    val effectiveRegParam = $(loss) match {
-      case SquaredError => $(regParam) / yStd
-      case Huber => $(regParam)
-    }
-    val effectiveL1RegParam = $(elasticNetParam) * effectiveRegParam
-    val effectiveL2RegParam = (1.0 - $(elasticNetParam)) * effectiveRegParam
-
-    val getFeaturesStd = (j: Int) => if (j >= 0 && j < numFeatures) featuresStd(j) else 0.0
-    val regularization = if (effectiveL2RegParam != 0.0) {
-      val shouldApply = (idx: Int) => idx >= 0 && idx < numFeatures
-      Some(new L2Regularization(effectiveL2RegParam, shouldApply,
-        if ($(standardization)) None else Some(getFeaturesStd)))
-    } else None
-
-    val optimizer = createOptimizer(effectiveRegParam, effectiveL1RegParam,
-      numFeatures, featuresStd)
-
-    val initialValues = $(loss) match {
-      case SquaredError =>
-        Vectors.zeros(numFeatures)
-      case Huber =>
-        val dim = if ($(fitIntercept)) numFeatures + 2 else numFeatures + 1
-        Vectors.dense(Array.fill(dim)(1.0))
-    }
-
-    val (parameters, objectiveHistory) =
-      trainImpl(instances, actualBlockSizeInMB, yMean, yStd,
-        featuresMean, featuresStd, initialValues, regularization, optimizer)
-
-    if (parameters == null) {
-      val msg = s"${optimizer.getClass.getName} failed."
-      instr.logError(msg)
-      throw new SparkException(msg)
-    }
-
-    val model = createModel(parameters, yMean, yStd, featuresMean, featuresStd)
-    // Handle possible missing or invalid prediction columns
-    val (summaryModel, predictionColName) = model.findSummaryModelAndPredictionCol()
-
-    val trainingSummary = new LinearRegressionTrainingSummary(
-      summaryModel.transform(dataset), predictionColName, $(labelCol), $(featuresCol),
-      model, Array(0.0), objectiveHistory)
-    model.setSummary(Some(trainingSummary))
-  }
-
-  private def trainWithNormal(
-      dataset: Dataset[_],
-      instr: Instrumentation): LinearRegressionModel = {
-    // oneDAL only support simple linear regression and ridge regression
-    val paramSupported = ($(regParam) == 0) || ($(regParam) != 0 && $(elasticNetParam) == 0)
-    val isPlatformSupported = Utils.checkClusterPlatformCompatibility(
-      dataset.sparkSession.sparkContext)
-    if (paramSupported && Utils.isOAPEnabled && isPlatformSupported) {
-      val executor_num = Utils.sparkExecutorNum(dataset.sparkSession.sparkContext)
-      val executor_cores = Utils.sparkExecutorCores()
-      logInfo(s"LinearRegressionDAL fit using $executor_num Executors")
-
-      val optimizer = new LinearRegressionDALImpl($(fitIntercept), $(regParam),
-        elasticNetParam = $(elasticNetParam), $(standardization), true,
-        executor_num, executor_cores)
-
-      // Return same model as WeightedLeastSquaresModel
-      val model = optimizer.train(dataset)
-
-      val lrModel = copyValues(
-        new LinearRegressionModel(uid, model.coefficients, model.intercept))
-
-      val (summaryModel, predictionColName) = lrModel.findSummaryModelAndPredictionCol()
-
-      val trainingSummary = new LinearRegressionTrainingSummary(
-        summaryModel.transform(dataset),
-        predictionColName,
-        $(labelCol),
-        $(featuresCol),
-        summaryModel,
-        model.diagInvAtWA.toArray,
-        model.objectiveHistory)
-
-      return lrModel.setSummary(Some(trainingSummary))
-    } else {
-      // For low dimensional data, WeightedLeastSquares is more efficient since the
-      // training algorithm only requires one pass through the data. (SPARK-10668)
-
-      val optimizer = new WeightedLeastSquares($(fitIntercept), $(regParam),
-        elasticNetParam = $(elasticNetParam), $(standardization), true,
-        solverType = WeightedLeastSquares.Auto, maxIter = $(maxIter), tol = $(tol))
-      val instances = extractInstances(dataset)
-        .setName("training instances")
-      val model = optimizer.fit(instances, instr = OptionalInstrumentation.create(instr))
-      // When it is trained by WeightedLeastSquares, training summary does not
-      // attach returned model.
-      val lrModel = copyValues(new LinearRegressionModel(uid, model.coefficients, model.intercept))
-      val (summaryModel, predictionColName) = lrModel.findSummaryModelAndPredictionCol()
-      val trainingSummary = new LinearRegressionTrainingSummary(
-        summaryModel.transform(dataset), predictionColName, $(labelCol), $(featuresCol),
-        summaryModel, model.diagInvAtWA.toArray, model.objectiveHistory)
-
-      lrModel.setSummary(Some(trainingSummary))
-    }
-  }
-
-  private def trainWithConstantLabel(
-      dataset: Dataset[_],
-      instr: Instrumentation,
-      numFeatures: Int,
-      yMean: Double): LinearRegressionModel = {
-    // If the rawYStd==0 and fitIntercept==true, then the intercept is yMean with
-    // zero coefficient; as a result, training is not needed.
-    // Also, if rawYStd==0 and yMean==0, all the coefficients are zero regardless of
-    // the fitIntercept.
-    if (yMean == 0.0) {
-      instr.logWarning(s"Mean and standard deviation of the label are zero, so the " +
-        s"coefficients and the intercept will all be zero; as a result, training is not " +
-        s"needed.")
-    } else {
-      instr.logWarning(s"The standard deviation of the label is zero, so the coefficients " +
-        s"will be zeros and the intercept will be the mean of the label; as a result, " +
-        s"training is not needed.")
-    }
-    val coefficients = Vectors.sparse(numFeatures, Seq.empty)
-    val intercept = yMean
-
-    val model = copyValues(new LinearRegressionModel(uid, coefficients, intercept))
-    // Handle possible missing or invalid prediction columns
-    val (summaryModel, predictionColName) = model.findSummaryModelAndPredictionCol()
-
-    val trainingSummary = new LinearRegressionTrainingSummary(
-      summaryModel.transform(dataset), predictionColName, $(labelCol), $(featuresCol),
-      model, Array(0.0), Array(0.0))
-
-    model.setSummary(Some(trainingSummary))
-  }
-
-  private def createOptimizer(
-      effectiveRegParam: Double,
-      effectiveL1RegParam: Double,
-      numFeatures: Int,
-      featuresStd: Array[Double]) = {
-    $(loss) match {
-      case SquaredError =>
-        if ($(elasticNetParam) == 0.0 || effectiveRegParam == 0.0) {
-          new BreezeLBFGS[BDV[Double]]($(maxIter), 10, $(tol))
-        } else {
-          val standardizationParam = $(standardization)
-          def effectiveL1RegFun = (index: Int) => {
-            if (standardizationParam) {
-              effectiveL1RegParam
-            } else {
-              // If `standardization` is false, we still standardize the data
-              // to improve the rate of convergence; as a result, we have to
-              // perform this reverse standardization by penalizing each component
-              // differently to get effectively the same objective function when
-              // the training dataset is not standardized.
-              if (featuresStd(index) != 0.0) effectiveL1RegParam / featuresStd(index) else 0.0
-            }
-          }
-          new BreezeOWLQN[Int, BDV[Double]]($(maxIter), 10, effectiveL1RegFun, $(tol))
-        }
-      case Huber =>
-        val dim = if ($(fitIntercept)) numFeatures + 2 else numFeatures + 1
-        val lowerBounds = BDV[Double](Array.fill(dim)(Double.MinValue))
-        // Optimize huber loss in space "\sigma > 0"
-        lowerBounds(dim - 1) = Double.MinPositiveValue
-        val upperBounds = BDV[Double](Array.fill(dim)(Double.MaxValue))
-        new BreezeLBFGSB(lowerBounds, upperBounds, $(maxIter), 10, $(tol))
-    }
-  }
-
-  private def trainImpl(
-      instances: RDD[Instance],
-      actualBlockSizeInMB: Double,
-      yMean: Double,
-      yStd: Double,
-      featuresMean: Array[Double],
-      featuresStd: Array[Double],
-      initialValues: Vector,
-      regularization: Option[L2Regularization],
-      optimizer: FirstOrderMinimizer[BDV[Double], DiffFunction[BDV[Double]]]) = {
-    val bcFeaturesMean = instances.context.broadcast(featuresMean)
-    val bcFeaturesStd = instances.context.broadcast(featuresStd)
-
-    val standardized = instances.mapPartitions { iter =>
-      val inverseStd = bcFeaturesStd.value.map { std => if (std != 0) 1.0 / std else 0.0 }
-      val func = StandardScalerModel.getTransformFunc(Array.empty, inverseStd, false, true)
-      iter.map { case Instance(label, weight, vec) => Instance(label, weight, func(vec)) }
-    }
-
-    val maxMemUsage = (actualBlockSizeInMB * 1024L * 1024L).ceil.toLong
-    val blocks = InstanceBlock.blokifyWithMaxMemUsage(standardized, maxMemUsage)
-      .persist(StorageLevel.MEMORY_AND_DISK)
-      .setName(s"training blocks (blockSizeInMB=$actualBlockSizeInMB)")
-
-    val costFun = $(loss) match {
-      case SquaredError =>
-        val getAggregatorFunc = new BlockLeastSquaresAggregator(yStd, yMean, $(fitIntercept),
-          bcFeaturesStd, bcFeaturesMean)(_)
-        new RDDLossFunction(blocks, getAggregatorFunc, regularization, $(aggregationDepth))
-      case Huber =>
-        val getAggregatorFunc = new BlockHuberAggregator($(fitIntercept), $(epsilon))(_)
-        new RDDLossFunction(blocks, getAggregatorFunc, regularization, $(aggregationDepth))
-    }
-
-    val states = optimizer.iterations(new CachedDiffFunction(costFun),
-      initialValues.asBreeze.toDenseVector)
-
-    /*
-       Note that in Linear Regression, the objective history (loss + regularization) returned
-       from optimizer is computed in the scaled space given by the following formula.
-       <blockquote>
-          $$
-          L &= 1/2n||\sum_i w_i(x_i - \bar{x_i}) / \hat{x_i} - (y - \bar{y}) / \hat{y}||^2
-               + regTerms \\
-          $$
-       </blockquote>
-     */
-    val arrayBuilder = mutable.ArrayBuilder.make[Double]
-    var state: optimizer.State = null
-    while (states.hasNext) {
-      state = states.next()
-      arrayBuilder += state.adjustedValue
-    }
-
-    blocks.unpersist()
-    bcFeaturesMean.destroy()
-    bcFeaturesStd.destroy()
-
-    (if (state == null) null else state.x.toArray, arrayBuilder.result)
-  }
-
-  private def createModel(
-      parameters: Array[Double],
-      yMean: Double,
-      yStd: Double,
-      featuresMean: Array[Double],
-      featuresStd: Array[Double]): LinearRegressionModel = {
-    val numFeatures = featuresStd.length
-    /*
-       The coefficients are trained in the scaled space; we're converting them back to
-       the original space.
-     */
-    val rawCoefficients = $(loss) match {
-      case SquaredError => parameters.clone()
-      case Huber => parameters.take(numFeatures)
-    }
-
-    var i = 0
-    val len = rawCoefficients.length
-    val multiplier = $(loss) match {
-      case SquaredError => yStd
-      case Huber => 1.0
-    }
-    while (i < len) {
-      rawCoefficients(i) *= { if (featuresStd(i) != 0.0) multiplier / featuresStd(i) else 0.0 }
-      i += 1
-    }
-
-    val intercept = if ($(fitIntercept)) {
-      $(loss) match {
-        case SquaredError =>
-          /*
-          The intercept of squared error in R's GLMNET is computed using closed form
-          after the coefficients are converged. See the following discussion for detail.
-          http://stats.stackexchange.com/questions/13617/how-is-the-intercept-computed-in-glmnet
-          */
-          yMean - dot(Vectors.dense(rawCoefficients), Vectors.dense(featuresMean))
-        case Huber => parameters(numFeatures)
-      }
-    } else 0.0
-
-    val scale = $(loss) match {
-      case SquaredError => 1.0
-      case Huber => parameters.last
-    }
-
-    val coefficients = Vectors.dense(rawCoefficients).compressed
-
-    copyValues(new LinearRegressionModel(uid, coefficients, intercept, scale))
+      dataset: Dataset[_]): LinearRegressionModel = {
+    val shim = LinearRegressionShim.create(uid)
+    shim.initShim(extractParamMap())
+    shim.train(dataset)
   }
 
   @Since("1.4.0")
@@ -711,397 +269,3 @@ object LinearRegression extends DefaultParamsReadable[LinearRegression] {
   /** Set of loss function names that LinearRegression supports. */
   private[regression] val supportedLosses = Array(SquaredError, Huber)
 }
-
-/**
- * Model produced by [[LinearRegression]].
- */
-@Since("1.3.0")
-class LinearRegressionModel private[ml] (
-    @Since("1.4.0") override val uid: String,
-    @Since("2.0.0") val coefficients: Vector,
-    @Since("1.3.0") val intercept: Double,
-    @Since("2.3.0") val scale: Double)
-  extends RegressionModel[Vector, LinearRegressionModel]
-  with LinearRegressionParams with GeneralMLWritable
-  with HasTrainingSummary[LinearRegressionTrainingSummary] {
-
-  private[ml] def this(uid: String, coefficients: Vector, intercept: Double) =
-    this(uid, coefficients, intercept, 1.0)
-
-  override val numFeatures: Int = coefficients.size
-
-  /**
-   * Gets summary (e.g. residuals, mse, r-squared ) of model on training set. An exception is
-   * thrown if `hasSummary` is false.
-   */
-  @Since("1.5.0")
-  override def summary: LinearRegressionTrainingSummary = super.summary
-
-  /**
-   * Evaluates the model on a test dataset.
-   *
-   * @param dataset Test dataset to evaluate model on.
-   */
-  @Since("2.0.0")
-  def evaluate(dataset: Dataset[_]): LinearRegressionSummary = {
-    // Handle possible missing or invalid prediction columns
-    val (summaryModel, predictionColName) = findSummaryModelAndPredictionCol()
-    new LinearRegressionSummary(summaryModel.transform(dataset), predictionColName,
-      $(labelCol), $(featuresCol), summaryModel, Array(0.0))
-  }
-
-  /**
-   * If the prediction column is set returns the current model and prediction column,
-   * otherwise generates a new column and sets it as the prediction column on a new copy
-   * of the current model.
-   */
-  private[regression] def findSummaryModelAndPredictionCol(): (LinearRegressionModel, String) = {
-    $(predictionCol) match {
-      case "" =>
-        val predictionColName = "prediction_" + java.util.UUID.randomUUID.toString
-        (copy(ParamMap.empty).setPredictionCol(predictionColName), predictionColName)
-      case p => (this, p)
-    }
-  }
-
-
-  override def predict(features: Vector): Double = {
-    dot(features, coefficients) + intercept
-  }
-
-  @Since("1.4.0")
-  override def copy(extra: ParamMap): LinearRegressionModel = {
-    val newModel = copyValues(new LinearRegressionModel(uid, coefficients, intercept), extra)
-    newModel.setSummary(trainingSummary).setParent(parent)
-  }
-
-  /**
-   * Returns a [[org.apache.spark.ml.util.GeneralMLWriter]] instance for this ML instance.
-   *
-   * For [[LinearRegressionModel]], this does NOT currently save the training [[summary]].
-   * An option to save [[summary]] may be added in the future.
-   *
-   * This also does not save the [[parent]] currently.
-   */
-  @Since("1.6.0")
-  override def write: GeneralMLWriter = new GeneralMLWriter(this)
-
-  @Since("3.0.0")
-  override def toString: String = {
-    s"LinearRegressionModel: uid=$uid, numFeatures=$numFeatures"
-  }
-}
-
-/** A writer for LinearRegression that handles the "internal" (or default) format */
-private class InternalLinearRegressionModelWriter
-  extends MLWriterFormat with MLFormatRegister {
-
-  override def format(): String = "internal"
-  override def stageName(): String = "org.apache.spark.ml.regression.LinearRegressionModel"
-
-  private case class Data(intercept: Double, coefficients: Vector, scale: Double)
-
-  override def write(path: String, sparkSession: SparkSession,
-    optionMap: mutable.Map[String, String], stage: PipelineStage): Unit = {
-    val instance = stage.asInstanceOf[LinearRegressionModel]
-    val sc = sparkSession.sparkContext
-    // Save metadata and Params
-    DefaultParamsWriter.saveMetadata(instance, path, sc)
-    // Save model data: intercept, coefficients, scale
-    val data = Data(instance.intercept, instance.coefficients, instance.scale)
-    val dataPath = new Path(path, "data").toString
-    sparkSession.createDataFrame(Seq(data)).repartition(1).write.parquet(dataPath)
-  }
-}
-
-/** A writer for LinearRegression that handles the "pmml" format */
-private class PMMLLinearRegressionModelWriter
-  extends MLWriterFormat with MLFormatRegister {
-
-  override def format(): String = "pmml"
-
-  override def stageName(): String = "org.apache.spark.ml.regression.LinearRegressionModel"
-
-  private case class Data(intercept: Double, coefficients: Vector)
-
-  override def write(path: String, sparkSession: SparkSession,
-    optionMap: mutable.Map[String, String], stage: PipelineStage): Unit = {
-    val sc = sparkSession.sparkContext
-    // Construct the MLLib model which knows how to write to PMML.
-    val instance = stage.asInstanceOf[LinearRegressionModel]
-    val oldModel = new OldLinearRegressionModel(instance.coefficients, instance.intercept)
-    // Save PMML
-    oldModel.toPMML(sc, path)
-  }
-}
-
-@Since("1.6.0")
-object LinearRegressionModel extends MLReadable[LinearRegressionModel] {
-
-  @Since("1.6.0")
-  override def read: MLReader[LinearRegressionModel] = new LinearRegressionModelReader
-
-  @Since("1.6.0")
-  override def load(path: String): LinearRegressionModel = super.load(path)
-
-  private class LinearRegressionModelReader extends MLReader[LinearRegressionModel] {
-
-    /** Checked against metadata when loading model */
-    private val className = classOf[LinearRegressionModel].getName
-
-    override def load(path: String): LinearRegressionModel = {
-      val metadata = DefaultParamsReader.loadMetadata(path, sc, className)
-
-      val dataPath = new Path(path, "data").toString
-      val data = sparkSession.read.format("parquet").load(dataPath)
-      val (majorVersion, minorVersion) = majorMinorVersion(metadata.sparkVersion)
-      val model = if (majorVersion < 2 || (majorVersion == 2 && minorVersion <= 2)) {
-        // Spark 2.2 and before
-        val Row(intercept: Double, coefficients: Vector) =
-          MLUtils.convertVectorColumnsToML(data, "coefficients")
-            .select("intercept", "coefficients")
-            .head()
-        new LinearRegressionModel(metadata.uid, coefficients, intercept)
-      } else {
-        // Spark 2.3 and later
-        val Row(intercept: Double, coefficients: Vector, scale: Double) =
-          data.select("intercept", "coefficients", "scale").head()
-        new LinearRegressionModel(metadata.uid, coefficients, intercept, scale)
-      }
-
-      metadata.getAndSetParams(model)
-      model
-    }
-  }
-}
-
-/**
- * Linear regression training results. Currently, the training summary ignores the
- * training weights except for the objective trace.
- *
- * @param predictions predictions output by the model's `transform` method.
- * @param objectiveHistory objective function (scaled loss + regularization) at each iteration.
- */
-@Since("1.5.0")
-class LinearRegressionTrainingSummary private[regression] (
-    predictions: DataFrame,
-    predictionCol: String,
-    labelCol: String,
-    featuresCol: String,
-    model: LinearRegressionModel,
-    diagInvAtWA: Array[Double],
-    val objectiveHistory: Array[Double])
-  extends LinearRegressionSummary(
-    predictions,
-    predictionCol,
-    labelCol,
-    featuresCol,
-    model,
-    diagInvAtWA) {
-
-  /**
-   * Number of training iterations until termination
-   *
-   * This value is only available when using the "l-bfgs" solver.
-   *
-   * @see `LinearRegression.solver`
-   */
-  @Since("1.5.0")
-  val totalIterations = {
-    assert(objectiveHistory.length > 0, s"objectiveHistory length should be greater than 1.")
-    objectiveHistory.length - 1
-  }
-}
-
-/**
- * Linear regression results evaluated on a dataset.
- *
- * @param predictions predictions output by the model's `transform` method.
- * @param predictionCol Field in "predictions" which gives the predicted value of the label at
- *                      each instance.
- * @param labelCol Field in "predictions" which gives the true label of each instance.
- * @param featuresCol Field in "predictions" which gives the features of each instance as a vector.
- */
-@Since("1.5.0")
-class LinearRegressionSummary private[regression] (
-    @transient val predictions: DataFrame,
-    val predictionCol: String,
-    val labelCol: String,
-    val featuresCol: String,
-    private val privateModel: LinearRegressionModel,
-    private val diagInvAtWA: Array[Double]) extends Serializable {
-
-  @transient private val metrics = {
-    val weightCol =
-      if (!privateModel.isDefined(privateModel.weightCol) || privateModel.getWeightCol.isEmpty) {
-        lit(1.0)
-      } else {
-        col(privateModel.getWeightCol).cast(DoubleType)
-      }
-
-    new RegressionMetrics(
-      predictions
-        .select(col(predictionCol), col(labelCol).cast(DoubleType), weightCol)
-        .rdd
-        .map { case Row(pred: Double, label: Double, weight: Double) => (pred, label, weight) },
-      !privateModel.getFitIntercept)
-  }
-
-  /**
-   * Returns the explained variance regression score.
-   * explainedVariance = 1 - variance(y - \hat{y}) / variance(y)
-   * Reference: <a href="http://en.wikipedia.org/wiki/Explained_variation">
-   * Wikipedia explain variation</a>
-   */
-  @Since("1.5.0")
-  val explainedVariance: Double = metrics.explainedVariance
-
-  /**
-   * Returns the mean absolute error, which is a risk function corresponding to the
-   * expected value of the absolute error loss or l1-norm loss.
-   */
-  @Since("1.5.0")
-  val meanAbsoluteError: Double = metrics.meanAbsoluteError
-
-  /**
-   * Returns the mean squared error, which is a risk function corresponding to the
-   * expected value of the squared error loss or quadratic loss.
-   */
-  @Since("1.5.0")
-  val meanSquaredError: Double = metrics.meanSquaredError
-
-  /**
-   * Returns the root mean squared error, which is defined as the square root of
-   * the mean squared error.
-   */
-  @Since("1.5.0")
-  val rootMeanSquaredError: Double = metrics.rootMeanSquaredError
-
-  /**
-   * Returns R^2^, the coefficient of determination.
-   * Reference: <a href="http://en.wikipedia.org/wiki/Coefficient_of_determination">
-   * Wikipedia coefficient of determination</a>
-   */
-  @Since("1.5.0")
-  val r2: Double = metrics.r2
-
-  /**
-   * Returns Adjusted R^2^, the adjusted coefficient of determination.
-   * Reference: <a href="https://en.wikipedia.org/wiki/Coefficient_of_determination#Adjusted_R2">
-   * Wikipedia coefficient of determination</a>
-   */
-  @Since("2.3.0")
-  val r2adj: Double = {
-    val interceptDOF = if (privateModel.getFitIntercept) 1 else 0
-    1 - (1 - r2) * (numInstances - interceptDOF) /
-      (numInstances - privateModel.coefficients.size - interceptDOF)
-  }
-
-  /** Residuals (label - predicted value) */
-  @Since("1.5.0")
-  @transient lazy val residuals: DataFrame = {
-    val t = udf { (pred: Double, label: Double) => label - pred }
-    predictions.select(t(col(predictionCol), col(labelCol)).as("residuals"))
-  }
-
-  /** Number of instances in DataFrame predictions */
-  lazy val numInstances: Long = metrics.count
-
-  /** Degrees of freedom */
-  @Since("2.2.0")
-  val degreesOfFreedom: Long = if (privateModel.getFitIntercept) {
-    numInstances - privateModel.coefficients.size - 1
-  } else {
-    numInstances - privateModel.coefficients.size
-  }
-
-  /**
-   * The weighted residuals, the usual residuals rescaled by
-   * the square root of the instance weights.
-   */
-  lazy val devianceResiduals: Array[Double] = {
-    val weighted =
-      if (!privateModel.isDefined(privateModel.weightCol) || privateModel.getWeightCol.isEmpty) {
-        lit(1.0)
-      } else {
-        sqrt(col(privateModel.getWeightCol))
-      }
-    val dr = predictions
-      .select(col(privateModel.getLabelCol).minus(col(privateModel.getPredictionCol))
-        .multiply(weighted).as("weightedResiduals"))
-      .select(min(col("weightedResiduals")).as("min"), max(col("weightedResiduals")).as("max"))
-      .first()
-    Array(dr.getDouble(0), dr.getDouble(1))
-  }
-
-  /**
-   * Standard error of estimated coefficients and intercept.
-   * This value is only available when using the "normal" solver.
-   *
-   * If `LinearRegression.fitIntercept` is set to true,
-   * then the last element returned corresponds to the intercept.
-   *
-   * @see `LinearRegression.solver`
-   */
-  lazy val coefficientStandardErrors: Array[Double] = {
-    if (diagInvAtWA.length == 1 && diagInvAtWA(0) == 0) {
-      throw new UnsupportedOperationException(
-        "No Std. Error of coefficients available for this LinearRegressionModel")
-    } else {
-      val rss =
-        if (!privateModel.isDefined(privateModel.weightCol) || privateModel.getWeightCol.isEmpty) {
-          meanSquaredError * numInstances
-        } else {
-          val t = udf { (pred: Double, label: Double, weight: Double) =>
-            math.pow(label - pred, 2.0) * weight }
-          predictions.select(t(col(privateModel.getPredictionCol), col(privateModel.getLabelCol),
-            col(privateModel.getWeightCol)).as("wse")).agg(sum(col("wse"))).first().getDouble(0)
-        }
-      val sigma2 = rss / degreesOfFreedom
-      diagInvAtWA.map(_ * sigma2).map(math.sqrt)
-    }
-  }
-
-  /**
-   * T-statistic of estimated coefficients and intercept.
-   * This value is only available when using the "normal" solver.
-   *
-   * If `LinearRegression.fitIntercept` is set to true,
-   * then the last element returned corresponds to the intercept.
-   *
-   * @see `LinearRegression.solver`
-   */
-  lazy val tValues: Array[Double] = {
-    if (diagInvAtWA.length == 1 && diagInvAtWA(0) == 0) {
-      throw new UnsupportedOperationException(
-        "No t-statistic available for this LinearRegressionModel")
-    } else {
-      val estimate = if (privateModel.getFitIntercept) {
-        Array.concat(privateModel.coefficients.toArray, Array(privateModel.intercept))
-      } else {
-        privateModel.coefficients.toArray
-      }
-      estimate.zip(coefficientStandardErrors).map { x => x._1 / x._2 }
-    }
-  }
-
-  /**
-   * Two-sided p-value of estimated coefficients and intercept.
-   * This value is only available when using the "normal" solver.
-   *
-   * If `LinearRegression.fitIntercept` is set to true,
-   * then the last element returned corresponds to the intercept.
-   *
-   * @see `LinearRegression.solver`
-   */
-  lazy val pValues: Array[Double] = {
-    if (diagInvAtWA.length == 1 && diagInvAtWA(0) == 0) {
-      throw new UnsupportedOperationException(
-        "No p-value available for this LinearRegressionModel")
-    } else {
-      tValues.map { x => 2.0 * (1.0 - StudentsT(degreesOfFreedom.toDouble).cdf(math.abs(x))) }
-    }
-  }
-
-}
-
